Check user and returning task repetition

+++ backend/routes/task.js
@@ -20,7 +20,35 @@ router.post("/create", verifyToken, async (req, res) => {
 router.get("/all", verifyToken, async (req, res) => {
   try {
     const task = await Task.findAll();
-    res.json(task);
+    var result = [...task];
+    task.forEach((t) => {
+      switch (t.repeat) {
+        case "Daily":
+          for (let i = 1; i <= 60; i++) {
+            const newDate = new Date(t.date);
+            newDate.setDate(newDate.getDate() + i);
+            result.push({ name: t.name, userId: t.userId, date: newDate, repeat: "None" });
+          }
+          break;
+        case "Weekly":
+          for (let i = 1; i <= 8; i++) {
+            const newDate = new Date(t.date);
+            newDate.setDate(newDate.getDate() + i * 7);
+            result.push({ name: t.name, userId: t.userId, date: newDate, repeat: "None" });
+          }
+          break;
+        case "Monthly":
+          for (let i = 1; i <= 3; i++) {
+            const newDate = new Date(t.date);
+            newDate.setMonth(newDate.getMonth() + i);
+            result.push({ name: t.name, userId: t.userId, date: newDate, repeat: "None" });
+          }
+          break;
+        case "None":
+          break;
+      }
+    });
+    res.json(result);
   } catch (error) {
     res.status(500).json({ error: "Failed to get all tasks." });
   }
@@ -30,6 +58,10 @@ router.get("/all", verifyToken, async (req, res) => {
 router.post("/findByUserIdAndDate", verifyToken, async (req, res) => {
   try {
     const { userId, date } = req.body;
+    const tokenUserId = req.userId;
+    if (tokenUserId !== userId) {
+      return res.status(403).json({ error: "Unauthorized access." });
+    }
     const dataStart = new Date(date).setHours(0, 0, 0, 0);
     const dataEnd = new Date(date).setHours(23, 59, 59, 59);
     const tasks = await Task.findAll({
@@ -51,6 +83,14 @@ router.post("/findByUserIdAndDate", verifyToken, async (req, res) => {
 router.delete("/delete/:id", verifyToken, async (req, res) => {
   try {
     const { id } = req.params;
+    const task = await Task.findOne({ where: { id } });
+    if (!task) {
+      return res.status(404).json({ error: "Task not found." });
+    }
+    const tokenUserId = req.userId;
+    if (tokenUserId !== task.userId) {
+      return res.status(403).json({ error: "Unauthorized access." });
+    }
     await Task.destroy({ where: { id } });
     res.sendStatus(204);
   } catch (error) {
@@ -64,6 +104,13 @@ router.put("/update/:id", verifyToken, async (req, res) => {
     const { id } = req.params;
     const { name, done, date } = req.body;
     const task = await Task.findOne({ where: { id } });
+    if (!task) {
+      return res.status(404).json({ error: "Task not found." });
+    }
+    const tokenUserId = req.userId;
+    if (tokenUserId !== task.userId) {
+      return res.status(403).json({ error: "Unauthorized access." });
+    }
     task.name = name;
     task.done = done;
     task.date = date;
